Added actions `select-value`, `add-value`, `remove-value`
These "value" actions- similar to their "item" counterparts- emit the selection _value_ as specified by the component's `optionValuePath`.

The aim is to provide an easier way to conform to the data-down-actions-up methodology.

Updating a target property can be done with a simple `{{mut}}` helper or other, without needing to create a specific handler to parse out. a specific value from a passed object.