Fix error on render when using `optionTemplate` + acceptance test
Couldn't get a unit test to work, but it still throws
`TypeError: Cannot set property '_elementInserted' of null`.

My guess is, that something between the jQuery event system and Ember is not
working in the isolated unit test, but it works in the browser and is now
properly tested with an acceptance test.